backend/requirements.txt
sqlalchemy
psycopg2-binary
python-dotenv
fastapi
uvicorn
requests
pydantic
watchdog
python-magic-bin
